Move template functions to separate foreign_globals.inline.hpp files
